# Move benchmark assigned err to global exported variable